requirements.txt
Django
gunicorn
psycopg2
redis
requests
six==1.14.0
websocket-client
django-storages
boto3==1.7.67
django-import-export
python-markdown-math
django-redis
ipython
pytz
django-websocket-redis
openpyxl
pyppeteer
raven
tablib
moratab
inotify
django-xff
pycairo
google-cloud-translate==2.0.1
